Address service clean up and moved dto to shared util module
