+++ nozama-api/Nozama_AddressService/src/main/java/in/nozama/address/service/model/mapper/AddressMapper.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 import org.springframework.stereotype.Component;
 
 import in.nozama.address.service.model.Address;
-import in.nozama.address.service.model.AddressResponse;
 import in.nozama.address.service.model.AddressType;
 import in.nozama.service.dto.AddressRequest;
+import in.nozama.service.model.AddressResponse;
 
 @Component
 public class AddressMapper {
@@ -19,10 +19,12 @@ public class AddressMapper {
 			to.setAddressId(address.getAddressId());
 			to.setAddress1(address.getAddress1());
 			to.setAddress2(address.getAddress2());
-			to.setAddressType(address.getAddressType().toString());
+			if (address.getAddressType() != null) {
+				to.setAddressType(address.getAddressType().toString());
+			}
 			to.setCity(address.getCity());
 			to.setCountry(address.getCountry());
-			to.setPinCode(address.getPinCode());
+			to.setPincode(address.getPincode());
 			to.setUserId(address.getUserId());
 			to.setState(address.getState());
 			to.setStreet(address.getStreet());
@@ -50,7 +52,7 @@ public class AddressMapper {
 			to.setCountry(from.getCountry());
 			to.setState(from.getState());
 			to.setStreet(from.getStreet());
-			to.setPinCode(from.getPinCode());
+			to.setPincode(from.getPincode());
 			to.setUserId(from.getUserId());
 		}
 		return to;
@@ -66,7 +68,7 @@ public class AddressMapper {
 			to.setCountry(from.getCountry());
 			to.setState(from.getState());
 			to.setStreet(from.getStreet());
-			to.setPinCode(from.getPinCode());
+			to.setPincode(from.getPincode());
 			to.setUserId(from.getUserId());
 		}
 		return to;
@@ -82,10 +84,30 @@ public class AddressMapper {
 			to.setCountry(from.getCountry());
 			to.setState(from.getState());
 			to.setStreet(from.getStreet());
-			to.setPinCode(from.getPinCode());
+			to.setPincode(from.getPincode());
 			to.setUserId(from.getUserId());
 		}
 		return to;
+	}
+
+	public AddressResponse map(Address from, AddressResponse to) {
+		return this.map(Optional.of(from));
+//		if (to == null) {
+//			to = new AddressResponse();
+//		}
+//		if (from != null) {
+//			to.setAddress1(from.getAddress1());
+//			to.setAddress2(from.getAddress2());
+//			to.setAddressId(from.getAddressId());
+//			to.setAddressType(from.getAddressType().getValue());
+//			to.setCity(from.getCity());
+//			to.setCountry(from.getCountry());
+//			to.setPinCode(from.getPinCode());
+//			to.setState(from.getState());
+//			to.setStreet(from.getStreet());
+//			to.setUserId(from.getUserId());
+//		}
+//		return to;
 	}
 
 }
