+++ nozama-api/Nozama_AddressService/src/main/java/in/nozama/address/service/service/AddressService.java
@@ -1,14 +1,14 @@
 package in.nozama.address.service.service;
 
 import in.nozama.address.service.exception.AddressNotFoundException;
-import in.nozama.address.service.model.AddressResponse;
 import in.nozama.service.dto.AddressRequest;
+import in.nozama.service.model.AddressResponse;
 
 public interface AddressService {
 
 	AddressResponse getAddressById(long addressId) throws AddressNotFoundException;
 
-	Long save(AddressRequest addressRequest);
+	AddressResponse save(AddressRequest addressRequest);
 
 	AddressResponse getAddressByUserId(Long userid);
