+++ nozama-api/Nozama_AddressService/src/main/resources/application-dev.yaml
@@ -24,7 +24,7 @@ spring:
     name: "nozama-addressservice"
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: "jdbc:mysql://${NOZAMA_APP_MYSQL_URL:localhost}:${NOZAMA_APP_MYSQL_PORT:3306}/${NOZAMA_APP_MYSQL_DBNAME:nozama_user_db}?createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true&useSSL=false"
+    url: "jdbc:mysql://${NOZAMA_APP_MYSQL_URL:localhost}:${NOZAMA_APP_MYSQL_PORT:3310}/${NOZAMA_APP_MYSQL_DBNAME:nozama_user_db}?createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true&useSSL=false"
     username: nozama_dba
     password: nozama19
