+++ nozama-api/Nozama_AddressService/src/test/java/in/nozama/address/service/NozamaAddressServiceApplicationTests.java
@@ -5,14 +5,8 @@ import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
-
-import in.nozama.address.service.exception.AddressNotFoundException;
-import in.nozama.address.service.model.AddressResponse;
-import in.nozama.address.service.service.AddressService;
-import in.nozama.address.service.service.AddressServiceImpl;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
