+++ nozama-api/Nozama_Shared_Lib/src/main/java/in/nozama/service/model/AddressResponse.java
@@ -1,13 +1,11 @@
-package in.nozama.address.service.model;
+package in.nozama.service.model;
 
 import java.io.Serializable;
 
 import com.fasterxml.jackson.annotation.JsonView;
 
 import in.nozama.service.dto.view.UserModelView;
-import in.nozama.service.model.ErrorResponse;
 import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -40,7 +38,7 @@ public class AddressResponse implements Serializable {
 
 	private String country;
 
-	private String pinCode;
+	private String pincode;
 	
 	private String addressType;
